+++ src/rpc/names.cpp
@@ -10,7 +10,6 @@
 #include "names/common.h"
 #include "names/main.h"
 #include "primitives/transaction.h"
-#include "rpc/safemode.h"
 #include "rpc/server.h"
 #include "script/names.h"
 #include "txmempool.h"
@@ -129,8 +128,6 @@ name_show (const JSONRPCRequest& request)
     throw JSONRPCError(RPC_CLIENT_IN_INITIAL_DOWNLOAD,
                        "Chimaera is downloading blocks...");
 
-  ObserveSafeMode ();
-
   const std::string nameStr = request.params[0].get_str ();
   const valtype name = ValtypeFromString (nameStr);
 
@@ -178,8 +175,6 @@ name_history (const JSONRPCRequest& request)
   if (IsInitialBlockDownload ())
     throw JSONRPCError(RPC_CLIENT_IN_INITIAL_DOWNLOAD,
                        "Chimaera is downloading blocks...");
-
-  ObserveSafeMode ();
 
   const std::string nameStr = request.params[0].get_str ();
   const valtype name = ValtypeFromString (nameStr);
@@ -239,8 +234,6 @@ name_scan (const JSONRPCRequest& request)
     throw JSONRPCError(RPC_CLIENT_IN_INITIAL_DOWNLOAD,
                        "Chimaera is downloading blocks...");
 
-  ObserveSafeMode ();
-
   valtype start;
   if (request.params.size () >= 1)
     start = ValtypeFromString (request.params[0].get_str ());
@@ -298,8 +291,6 @@ name_filter (const JSONRPCRequest& request)
   if (IsInitialBlockDownload ())
     throw JSONRPCError(RPC_CLIENT_IN_INITIAL_DOWNLOAD,
                        "Chimaera is downloading blocks...");
-
-  ObserveSafeMode ();
 
   /* ********************** */
   /* Interpret parameters.  */