+++ src/wallet/rpcnames.cpp
@@ -11,7 +11,6 @@
 #include "primitives/transaction.h"
 #include "random.h"
 #include "rpc/mining.h"
-#include "rpc/safemode.h"
 #include "rpc/server.h"
 #include "script/names.h"
 #include "txmempool.h"
@@ -49,8 +48,6 @@ name_list (const JSONRPCRequest& request)
       );
 
   RPCTypeCheck (request.params, {UniValue::VSTR});
-
-  ObserveSafeMode ();
 
   valtype nameFilter;
   if (request.params.size () == 1)
@@ -148,8 +145,6 @@ name_register (const JSONRPCRequest& request)
 
   RPCTypeCheck (request.params,
                 {UniValue::VSTR, UniValue::VSTR, UniValue::VSTR});
-
-  ObserveSafeMode ();
 
   const std::string nameStr = request.params[0].get_str ();
   const valtype name = ValtypeFromString (nameStr);
@@ -250,8 +245,6 @@ name_update (const JSONRPCRequest& request)
 
   RPCTypeCheck (request.params,
                 {UniValue::VSTR, UniValue::VSTR, UniValue::VSTR});
-
-  ObserveSafeMode ();
 
   const std::string nameStr = request.params[0].get_str ();
   const valtype name = ValtypeFromString (nameStr);
@@ -374,7 +367,6 @@ sendtoname (const JSONRPCRequest& request)
     throw JSONRPCError(RPC_CLIENT_IN_INITIAL_DOWNLOAD,
                        "Chimaera is downloading blocks...");
 
-  ObserveSafeMode ();
   LOCK2 (cs_main, pwallet->cs_wallet);
 
   const std::string nameStr = request.params[0].get_str ();